Move pkg/subctl/operator/submarinercr under pkg
