+++ internal/show/versions.go
@@ -26,7 +26,7 @@ import (
 	"github.com/submariner-io/admiral/pkg/reporter"
 	"github.com/submariner-io/subctl/internal/constants"
 	"github.com/submariner-io/subctl/pkg/cluster"
-	"github.com/submariner-io/subctl/pkg/subctl/operator/submarinercr"
+	"github.com/submariner-io/subctl/pkg/submarinercr"
 	"github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	submarinerclientset "github.com/submariner-io/submariner-operator/pkg/client/clientset/versioned"
 	"github.com/submariner-io/submariner-operator/pkg/images"
