+++ pkg/deploy/submariner.go
@@ -26,7 +26,7 @@ import (
 	"github.com/submariner-io/subctl/internal/constants"
 	"github.com/submariner-io/subctl/pkg/broker"
 	"github.com/submariner-io/subctl/pkg/secret"
-	"github.com/submariner-io/subctl/pkg/subctl/operator/submarinercr"
+	"github.com/submariner-io/subctl/pkg/submarinercr"
 	submariner "github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	"github.com/submariner-io/submariner-operator/pkg/client"
 	"github.com/submariner-io/submariner-operator/pkg/discovery/globalnet"
