+++ pkg/uninstall/all.go
@@ -28,7 +28,7 @@ import (
 	"github.com/submariner-io/admiral/pkg/util"
 	"github.com/submariner-io/subctl/internal/constants"
 	"github.com/submariner-io/subctl/pkg/subctl/operator/servicediscoverycr"
-	"github.com/submariner-io/subctl/pkg/subctl/operator/submarinercr"
+	"github.com/submariner-io/subctl/pkg/submarinercr"
 	"github.com/submariner-io/submariner-operator/pkg/client"
 	"github.com/submariner-io/submariner-operator/pkg/names"
 	corev1 "k8s.io/api/core/v1"
